Move save state progress dialog to ui/dialogs
