+++ src/ui/dialogs/SaveTableWin32Visitor.h
@@ -2,7 +2,7 @@
 
 #include "core/stdafx.h"
 
-#include "SaveTableVisitor.h"
+#include "parts/SaveTableVisitor.h"
 
 
 /**
